requirements.txt
appjar==0.070
python-i18n==0.3.1
pyyaml==3.12
sshtunnel==0.1.2
urllib==1.21.1
configparser==3.5.0